Accept content of text field when keyboard hides
When the text field for the destination name loses focus, its content
should be accepted. A text field loses focus not only when the user hits
the enter key, but also when the user touches another responsive
element, and when the user closes the sheet including the text field. In
those cases the user would expect the content to be accepted.